% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{decryptSound}
\alias{decryptSound}
\title{Break captchas sound}
\usage{
decryptSound(mp3_encoded, guid)
}
\arguments{
\item{mp3_encoded}{.}

\item{guid}{}
}
\description{
Given one mp3 and one guid code.
}
